refactor: move stream_converter logic into step_adaptor.process_chunks()
Consolidate AG-UI event generation into DRAgentNestedReasoningStepAdaptor.
Text streaming delegates to _handle_llm_primary_function via synthetic
IntermediateStepPayload. Tool call streaming uses new _handle_tool_call_chunk.
Delete standalone stream_converter module.
