+++ src/datarobot_genai/dragent/plugins/per_user_tool_calling_agent.py
@@ -22,8 +22,8 @@ that per-user function groups can be used while still benefiting from OpenAI-sty
 structured tool calling (``bind_tools``).
 
 NAT 1.6 added a ``stream_fn`` that yields ``ChatResponseChunk``.  We wrap it
-to produce ``DRAgentEventResponse`` with valid AG-UI event sequences instead.
-See ``stream_converter.py`` for the conversion logic.
+using ``DRAgentNestedReasoningStepAdaptor.process_chunks()`` to produce
+``DRAgentEventResponse`` with valid AG-UI event sequences.
 """
 
 from collections.abc import AsyncGenerator
@@ -34,11 +34,12 @@ from nat.cli.register_workflow import register_per_user_function
 from nat.data_models.api_server import ChatRequest
 from nat.data_models.api_server import ChatRequestOrMessage
 from nat.data_models.api_server import ChatResponse
+from nat.data_models.step_adaptor import StepAdaptorConfig
 from nat.plugins.langchain.agent.tool_calling_agent.register import ToolCallAgentWorkflowConfig
 from nat.plugins.langchain.agent.tool_calling_agent.register import tool_calling_agent_workflow
 
 from datarobot_genai.dragent.frontends.response import DRAgentEventResponse
-from datarobot_genai.dragent.plugins.stream_converter import convert_chunks_to_ag_ui_events
+from datarobot_genai.dragent.frontends.step_adaptor import DRAgentNestedReasoningStepAdaptor
 
 
 class PerUserToolCallAgentWorkflowConfig(
@@ -69,9 +70,8 @@ async def _per_user_tool_calling_agent(
         async def wrapped_stream(
             chat_request_or_message: ChatRequestOrMessage,
         ) -> AsyncGenerator[DRAgentEventResponse, None]:
-            async for event in convert_chunks_to_ag_ui_events(
-                original_stream_fn(chat_request_or_message)
-            ):
+            adaptor = DRAgentNestedReasoningStepAdaptor(StepAdaptorConfig())
+            async for event in adaptor.process_chunks(original_stream_fn(chat_request_or_message)):
                 yield event
 
         yield FunctionInfo.create(
