+++ tests/dragent/frontends/test_step_adaptor.py
@@ -12,8 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import datetime
 import json
 import uuid
+from collections.abc import AsyncGenerator
 from types import SimpleNamespace
 from unittest import mock
 
@@ -39,6 +41,11 @@ from nat.builder.context import IntermediateStep
 from nat.builder.context import IntermediateStepPayload
 from nat.builder.context import IntermediateStepType
 from nat.builder.context import InvocationNode
+from nat.data_models.api_server import ChatResponseChunk
+from nat.data_models.api_server import ChatResponseChunkChoice
+from nat.data_models.api_server import ChoiceDelta
+from nat.data_models.api_server import ChoiceDeltaToolCall
+from nat.data_models.api_server import ChoiceDeltaToolCallFunction
 from nat.data_models.intermediate_step import StreamEventData
 from nat.data_models.intermediate_step import TokenUsageBaseModel
 from nat.data_models.intermediate_step import TraceMetadata
@@ -774,3 +781,205 @@ class TestHandleToolArgsEncoding:
         response = step_adaptor.process(step)
         args_event = next(e for e in response.events if isinstance(e, ToolCallArgsEvent))
         assert args_event.delta == "{}"
+
+
+# --- Helpers for process_chunks tests ---
+
+
+def _make_chunk(
+    content: str | None = None,
+    tool_calls: list[ChoiceDeltaToolCall] | None = None,
+    chunk_id: str = "chunk-1",
+) -> ChatResponseChunk:
+    return ChatResponseChunk(
+        id=chunk_id,
+        choices=[
+            ChatResponseChunkChoice(
+                index=0,
+                delta=ChoiceDelta(content=content, tool_calls=tool_calls),
+            )
+        ],
+        created=datetime.datetime.now(datetime.UTC),
+    )
+
+
+async def _async_iter(*items) -> AsyncGenerator[ChatResponseChunk, None]:
+    for item in items:
+        yield item
+
+
+async def _collect(gen: AsyncGenerator) -> list:
+    return [item async for item in gen]
+
+
+def _flat_events(responses):
+    events = []
+    for resp in responses:
+        events.extend(resp.events)
+    return events
+
+
+def _make_adaptor():
+    return DRAgentNestedReasoningStepAdaptor(StepAdaptorConfig())
+
+
+class TestProcessChunksTextMessage:
+    @pytest.mark.asyncio
+    async def test_single_text_chunk_produces_start_content_end(self):
+        adaptor = _make_adaptor()
+        chunk = _make_chunk(content="Hello")
+        responses = await _collect(adaptor.process_chunks(_async_iter(chunk)))
+        events = _flat_events(responses)
+
+        assert len(events) == 3
+        assert isinstance(events[0], TextMessageStartEvent)
+        assert events[0].message_id == "chunk-1"
+        assert isinstance(events[1], TextMessageContentEvent)
+        assert events[1].delta == "Hello"
+        assert isinstance(events[2], TextMessageEndEvent)
+
+    @pytest.mark.asyncio
+    async def test_multiple_text_chunks_share_single_start_end(self):
+        adaptor = _make_adaptor()
+        chunks = [
+            _make_chunk(content="Hello ", chunk_id="c1"),
+            _make_chunk(content="world", chunk_id="c1"),
+        ]
+        responses = await _collect(adaptor.process_chunks(_async_iter(*chunks)))
+        events = _flat_events(responses)
+
+        start_events = [e for e in events if isinstance(e, TextMessageStartEvent)]
+        content_events = [e for e in events if isinstance(e, TextMessageContentEvent)]
+        end_events = [e for e in events if isinstance(e, TextMessageEndEvent)]
+
+        assert len(start_events) == 1
+        assert len(content_events) == 2
+        assert content_events[0].delta == "Hello "
+        assert content_events[1].delta == "world"
+        assert len(end_events) == 1
+
+
+class TestProcessChunksToolCall:
+    @pytest.mark.asyncio
+    async def test_single_tool_call_produces_start_args_end(self):
+        adaptor = _make_adaptor()
+        chunk = _make_chunk(
+            tool_calls=[
+                ChoiceDeltaToolCall(
+                    index=0,
+                    id="tc-1",
+                    function=ChoiceDeltaToolCallFunction(name="get_weather", arguments='{"loc":'),
+                )
+            ]
+        )
+        responses = await _collect(adaptor.process_chunks(_async_iter(chunk)))
+        events = _flat_events(responses)
+
+        assert len(events) == 3
+        assert isinstance(events[0], ToolCallStartEvent)
+        assert events[0].tool_call_id == "tc-1"
+        assert events[0].tool_call_name == "get_weather"
+        assert isinstance(events[1], ToolCallArgsEvent)
+        assert events[1].delta == '{"loc":'
+        assert isinstance(events[2], ToolCallEndEvent)
+
+    @pytest.mark.asyncio
+    async def test_tool_call_followup_chunks_use_index_lookup(self):
+        adaptor = _make_adaptor()
+        first = _make_chunk(
+            tool_calls=[
+                ChoiceDeltaToolCall(
+                    index=0,
+                    id="tc-1",
+                    function=ChoiceDeltaToolCallFunction(name="search", arguments='{"q":'),
+                )
+            ]
+        )
+        followup = _make_chunk(
+            tool_calls=[
+                ChoiceDeltaToolCall(
+                    index=0,
+                    id=None,
+                    function=ChoiceDeltaToolCallFunction(arguments='"hello"}'),
+                )
+            ]
+        )
+        responses = await _collect(adaptor.process_chunks(_async_iter(first, followup)))
+        events = _flat_events(responses)
+
+        args_events = [e for e in events if isinstance(e, ToolCallArgsEvent)]
+        assert len(args_events) == 2
+        assert args_events[0].tool_call_id == "tc-1"
+        assert args_events[1].tool_call_id == "tc-1"
+
+    @pytest.mark.asyncio
+    async def test_multiple_parallel_tool_calls(self):
+        adaptor = _make_adaptor()
+        chunk = _make_chunk(
+            tool_calls=[
+                ChoiceDeltaToolCall(
+                    index=0,
+                    id="tc-1",
+                    function=ChoiceDeltaToolCallFunction(name="tool_a", arguments="{}"),
+                ),
+                ChoiceDeltaToolCall(
+                    index=1,
+                    id="tc-2",
+                    function=ChoiceDeltaToolCallFunction(name="tool_b", arguments="{}"),
+                ),
+            ]
+        )
+        responses = await _collect(adaptor.process_chunks(_async_iter(chunk)))
+        events = _flat_events(responses)
+
+        start_events = [e for e in events if isinstance(e, ToolCallStartEvent)]
+        end_events = [e for e in events if isinstance(e, ToolCallEndEvent)]
+        assert len(start_events) == 2
+        assert {e.tool_call_id for e in start_events} == {"tc-1", "tc-2"}
+        assert len(end_events) == 2
+
+
+class TestProcessChunksEdgeCases:
+    @pytest.mark.asyncio
+    async def test_empty_stream_yields_nothing(self):
+        adaptor = _make_adaptor()
+        responses = await _collect(adaptor.process_chunks(_async_iter()))
+        assert responses == []
+
+    @pytest.mark.asyncio
+    async def test_chunk_with_no_content_and_no_tool_calls_is_skipped(self):
+        adaptor = _make_adaptor()
+        chunk = _make_chunk(content=None, tool_calls=None)
+        responses = await _collect(adaptor.process_chunks(_async_iter(chunk)))
+        assert responses == []
+
+
+class TestProcessChunksErrorHandling:
+    @pytest.mark.asyncio
+    async def test_end_events_emitted_on_upstream_exception(self):
+        adaptor = _make_adaptor()
+
+        async def _failing_gen():
+            yield _make_chunk(content="Hello")
+            raise RuntimeError("upstream error")
+
+        responses = []
+        with pytest.raises(RuntimeError, match="upstream error"):
+            async for resp in adaptor.process_chunks(_failing_gen()):
+                responses.append(resp)
+
+        events = _flat_events(responses)
+        assert any(isinstance(e, TextMessageStartEvent) for e in events)
+        assert any(isinstance(e, TextMessageEndEvent) for e in events)
+
+    @pytest.mark.asyncio
+    async def test_aclose_during_stream_does_not_raise(self):
+        adaptor = _make_adaptor()
+
+        async def _slow_gen():
+            yield _make_chunk(content="Hello")
+            yield _make_chunk(content=" world")
+
+        gen = adaptor.process_chunks(_slow_gen())
+        await gen.__anext__()
+        await gen.aclose()
